feat: 초기 member 및 post 데이터 초기화 로직
